+++ src/main/scala/lerna/akka/entityreplication/ClusterReplicationSettings.scala
@@ -41,6 +41,8 @@ trait ClusterReplicationSettings {
 
   def withDisabledShards(disabledShards: Set[String]): ClusterReplicationSettings
 
+  def withStickyLeaders(stickyLeaders: Map[String, String]): ClusterReplicationSettings
+
   def withRaftJournalPluginId(pluginId: String): ClusterReplicationSettings
 
   def withRaftSnapshotPluginId(pluginId: String): ClusterReplicationSettings