+++ src/main/scala/lerna/akka/entityreplication/internal/ClusterReplicationSettingsImpl.scala
@@ -21,6 +21,9 @@ private[entityreplication] final case class ClusterReplicationSettingsImpl(
   override def withDisabledShards(disabledShards: Set[String]): ClusterReplicationSettings =
     copy(raftSettings = raftSettings.withDisabledShards(disabledShards))
 
+  override def withStickyLeaders(stickyLeaders: Map[String, String]): ClusterReplicationSettingsImpl =
+    copy(raftSettings = raftSettings.withStickyLeaders(stickyLeaders))
+
   override def withRaftJournalPluginId(pluginId: String): ClusterReplicationSettingsImpl =
     copy(raftSettings = raftSettings.withJournalPluginId(pluginId))