+++ src/main/scala/lerna/akka/entityreplication/raft/RaftSettings.scala
@@ -14,6 +14,8 @@ trait RaftSettings {
 
   def electionTimeout: FiniteDuration
 
+  def stickyLeaders: Map[String, String]
+
   private[raft] def randomizedElectionTimeout(): FiniteDuration
 
   def heartbeatInterval: FiniteDuration
@@ -86,6 +88,8 @@ trait RaftSettings {
   def eventSourcedSnapshotEvery: Int
 
   private[entityreplication] def withDisabledShards(disabledShards: Set[String]): RaftSettings
+
+  private[entityreplication] def withStickyLeaders(stickyLeaders: Map[String, String]): RaftSettings
 
   private[entityreplication] def withJournalPluginId(pluginId: String): RaftSettings