+++ src/main/scala/lerna/akka/entityreplication/raft/RaftSettingsImpl.scala
@@ -11,6 +11,7 @@ import scala.util.Random
 private[entityreplication] final case class RaftSettingsImpl(
     config: Config,
     electionTimeout: FiniteDuration,
+    stickyLeaders: Map[String, String],
     heartbeatInterval: FiniteDuration,
     electionTimeoutMin: Duration,
     multiRaftRoles: Set[String],
@@ -59,6 +60,9 @@ private[entityreplication] final case class RaftSettingsImpl(
   override private[entityreplication] def withDisabledShards(disabledShards: Set[String]): RaftSettings =
     copy(disabledShards = disabledShards)
 
+  override private[entityreplication] def withStickyLeaders(stickyLeaders: Map[String, String]) =
+    copy(stickyLeaders = stickyLeaders)
+
   override private[entityreplication] def withJournalPluginId(pluginId: String): RaftSettings =
     copy(journalPluginId = pluginId)
 
@@ -83,6 +87,8 @@ private[entityreplication] object RaftSettingsImpl {
     val config: Config = root.getConfig("lerna.akka.entityreplication.raft")
 
     val electionTimeout: FiniteDuration = config.getDuration("election-timeout").toScala
+
+    val stickyLeaders: Map[String, String] = Map.empty
 
     val heartbeatInterval: FiniteDuration = config.getDuration("heartbeat-interval").toScala
 
@@ -231,6 +237,7 @@ private[entityreplication] object RaftSettingsImpl {
     RaftSettingsImpl(
       config = config,
       electionTimeout = electionTimeout,
+      stickyLeaders = stickyLeaders,
       heartbeatInterval = heartbeatInterval,
       electionTimeoutMin = electionTimeoutMin,
       multiRaftRoles = multiRaftRoles,